feat(ddia): aplicar conceitos de transações e consistência em interações e vídeo
  Aplica conceitos do livro Designing Data-Intensive Applications (caps. 7, 8 e 9)
  no domínio de likes, comentários e publicação de vídeo, com docs e testes.

  - Likes (cap. 7): toggle idempotente tratando corrida na constraint única,
    evitando HTTP 500 em double-tap/retry
  - Comentários (cap. 8): ordenação determinística por date+id do servidor;
    contagem via COUNT/GROUP BY no banco em vez de carregar linhas
  - Vídeo (cap. 9): publicação do evento só após o commit via
    @TransactionalEventListener(AFTER_COMMIT) + REQUIRES_NEW, corrigindo
    cross-channel timing latente no caminho SQS
  - Docs: 3 novos conceito-e-aplicacao.md em conceitos_livro_Designing-Data-Intensive_Applications
  - Testes: corrida idempotente em likes e dispatcher de evento de vídeo

+++ src/main/java/com/dvFabricio/VidaLongaFlix/controllers/analytics/AnalyticsCommentController.java
@@ -17,8 +17,8 @@ public class AnalyticsCommentController {
     private final CommentService commentService;
 
     @GetMapping("/quantidade/video/{videoId}")
-    public ResponseEntity<Integer> getCommentCountByVideo(@PathVariable UUID videoId) {
-        int count = commentService.getCommentCountByVideo(videoId);
+    public ResponseEntity<Long> getCommentCountByVideo(@PathVariable UUID videoId) {
+        long count = commentService.getCommentCountByVideo(videoId);
         return ResponseEntity.ok(count);
     }
